otaclient_common: split typing related from __init__ to typing module
