+++ src/otaclient_common/downloader.py
@@ -55,6 +55,8 @@ from ota_proxy import OTAFileCacheControl
 from otaclient.app.configs import config as cfg
 from otaclient_common import copy_callable_typehint
 from otaclient_common.common import wait_with_backoff
+from otaclient_common.typing import copy_callable_typehint
+
 
 logger = logging.getLogger(__name__)
